pyproject.toml
[project]
name = "flowchart-agent"
version = "0.1.0"
description = "A conversational AI agent that follows Mermaid flowcharts using Google ADK"
requires-python = ">=3.11"
dependencies = [
    "google-adk",
    "aiosqlite",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["flowchart_agent"]
